requirements.txt
discord
requests
python-dotenv